+++ misc/client-extensions/ClientExtensions/ClientExtensions.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "ClientDB.h"
+#include "CustomDBCMgr/ClientDB.h"
 #include "Character/CharacterFixes.h"
 #include "Tooltip/SpellTooltipExtensions.h"
