+++ misc/client-extensions/ClientExtensions/CustomDBCMgr/ClientDB.h
